fix(command-bar): open highlighted result on Enter for plain queries
In InPlace mode, Enter opened the typed text as a URL/search even when a
real result was highlighted (e.g. typing "terminal" searched Google
instead of opening the highlighted Terminal page). Clicking worked because
it executes the item directly.

Guard the typed-query shortcut with looks_like_url so it only fires for
actual URLs/domains; plain search terms fall through to executing the
selected result, matching click behavior. Move the decision fn to
vmux_command::event so its tests run in native CI.
